無限ストレージ/RAG チャンキング/requirements.txt
langchain>=0.1.0
tiktoken>=0.5.1
qrcode>=7.4.2
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
